Restructure solution tree and add build support for solution
